bug fix
Can't call setState on a component that is not yet mounted. This is a
no-op, but it might indicate a bug in your application. Instead, assign
to `this.state` directly or define a `state = {};` class property with
the desired state in the Calculator component.

Export class Component in another file
